Refactor Experience section
