+++ src/components/Experience/Job/Job.module.css
@@ -1,26 +1,3 @@
-/* .jobSection {
-  display: flex;
-  flex-direction: column;
-}
-
-.jobHeader {
-  display: flex;
-  justify-content: space-between;
-  font-weight: 500;
-  font-size: large;
-}
-
-.jobSubHeader {
-  display: flex;
-  justify-content: space-between;
-  color: var(--frost-2);
-}
-
-.jobMain {
-  display: flex;
-  flex-direction: column;
-} */
-
 .card {
   display: flex;
   gap: 20px;
