--- src/components/Experience/jobs.ts
@@ -1,40 +1,0 @@
-import { JobType } from "./Experience";
-
-export const jobs: JobType[] = [
-  {
-    id: 3,
-    title: "Analyst",
-    location: "Hatfield, UK",
-    company: "Ocado",
-    employmentPeriod: { startDate: "2022-10-01", endDate: "" },
-    workHighlights: [
-      "Lead Supply Chain Analytics projects for Christmas 2023 for both Ocado and Morrisons.",
-      "Create/Maintain/Refactor reports across platforms including BigQuery, Oracle, Tableau, Google Sheets, and Excel.",
-      "Champion a culture of code review and comprehensive documentation within the team.",
-      "Coach fellow team members and establish code conventions among the team.",
-    ],
-  },
-  {
-    id: 2,
-    title: "Supply Chain Manager",
-    location: "London, UK",
-    company: "Samarkand Global",
-    employmentPeriod: { startDate: "2021-04-01", endDate: "2022-09-01" },
-    workHighlights: [
-      "Designed and developed databases using BigQuery. Collected data from spreadsheets into databases for forthcoming analyses. ",
-      "Delivered data analysis and actionable insights to empower senior management in their data-oriented decision-making processes.",
-      "As a power user, closely collaborating with the Tech team to optimise the existing ERP system and introduce new features.",
-    ],
-  },
-  {
-    id: 1,
-    title: "Supply Chain Specialist",
-    location: "London, UK",
-    company: "Samarkand Global",
-    employmentPeriod: { startDate: "2019-09-01", endDate: "2021-03-01" },
-    workHighlights: [
-      "Created Python automation tools for diverse business functions, replacing manual tasks and slashing time requirements from weeks to minutes. ",
-      "Quickly emerged as a key ERP contributor and collaborator, leading training sessions and authoring a comprehensive user handbook for the company within just 6 months.",
-    ],
-  },
-];
