+++ src/styles/Home.module.css
@@ -1,0 +1,5 @@
+.sections {
+  display: flex;
+  flex-direction: column;
+  gap: 100px;
+}
